Refactor imports for clarity and modularity
Updated import statements across several modules to use more explicit and modular paths, improving code clarity and maintainability. Removed unnecessary blank lines in curriculum.py and test_history_callback.py. No functional changes were made.
